requirements.txt
-i https://pypi.org/simple
airportsdata==20230408
apscheduler==3.10.1
async-generator==1.10 ; python_version >= '3.5'
attrs==22.2.0 ; python_version >= '3.6'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
cached-property==1.5.2
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
comtypes==1.1.14 ; sys_platform == 'win32'
convertdate==2.4.0 ; python_version >= '3.7' and python_version < '4'
cryptography==39.0.2 ; python_version >= '3.6'
decorator==5.1.1 ; python_version >= '3.5'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
docutils==0.19 ; python_version >= '3.7'
et-xmlfile==1.1.0 ; python_version >= '3.6'
exceptiongroup==1.1.1 ; python_version < '3.11'
exchangelib==4.9.0 ; python_version >= '3.7'
fire==0.4.0
fonttools==4.39.3 ; python_version >= '3.8'
fpdf2==2.7.3 ; python_version >= '3.7'
furl==2.1.3
graphviz==0.13.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
h11==0.14.0 ; python_version >= '3.7'
hijri-converter==2.2.4 ; python_version >= '3.6'
holidays==0.21.13 ; python_version >= '3.7'
htmldocx==0.0.6 ; python_version >= '3.6'
hubspot-api-client==4.0.6 ; python_version >= '3.5'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==4.13.0 ; python_version >= '3.7'
isodate==0.6.1
java-access-bridge-wrapper==0.9.7 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
jsonpath-ng==1.5.3
jsonschema==4.17.3 ; python_version >= '3.7'
korean-lunar-calendar==0.3.1
lxml==4.9.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3 ; python_version >= '3.7' and python_version < '4.0'
o365==2.0.26 ; python_version >= '3.4'
oauthlib==3.2.2 ; python_version >= '3.6'
openpyxl==3.1.2 ; python_version >= '3.6'
orderedmultidict==1.0.1
outcome==1.2.0 ; python_version >= '3.7'
packaging==23.0 ; python_version >= '3.7'
pdfminer.six==20221105 ; python_version >= '3.6'
pendulum==2.1.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pillow==9.5.0 ; python_version >= '3.7'
platformdirs==3.2.0 ; python_version >= '3.7'
ply==3.11
psutil==5.9.4 ; sys_platform == 'win32'
pycparser==2.21
pygments==2.14.0 ; python_version >= '3.6'
pyjwt==2.6.0 ; python_version >= '3.7'
pymeeus==0.5.12
pynput-robocorp-fork==5.0.0
pyotp==2.8.0 ; python_version >= '3.7'
pypdf==3.7.1 ; python_version >= '3.6'
pyperclip==1.8.2
pyrsistent==0.19.3 ; python_version >= '3.7'
pysocks==1.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyspnego==0.8.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-docx==0.8.11
python-dotenv==1.0.0 ; python_version >= '3.8'
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pywin32==303 ; python_full_version != '3.7.6' and python_full_version != '3.8.1' and sys_platform == 'win32'
pywinauto==0.6.8 ; python_full_version != '3.7.6' and python_full_version != '3.8.1' and sys_platform == 'win32'
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-file==1.5.1
requests-ntlm==1.2.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==0.10.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
robotframework==5.0.1 ; python_version >= '3.6'
robotframework-pythonlibcore==4.1.2 ; python_version >= '3.7' and python_version < '4'
robotframework-requests==0.9.4
robotframework-sapguilibrary==1.1 ; sys_platform == 'win32'
robotframework-seleniumlibrary==6.0.0 ; python_version >= '3.6' and python_version < '4'
robotframework-seleniumtestability==2.1.0
rpaframework==22.5.1
rpaframework-core==11.0.0 ; python_version >= '3.7' and python_version < '4.0'
rpaframework-pdf==7.1.5 ; python_version >= '3.7' and python_version < '4.0'
rpaframework-windows==7.3.2 ; sys_platform == 'win32'
selenium==4.5.0
setuptools==67.6.1 ; python_version >= '3.7'
simple-salesforce==1.12.3
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smartsheet-python-sdk==3.0.0
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
soupsieve==2.4 ; python_version >= '3.7'
stringcase==1.2.0
tenacity==8.2.2 ; python_version >= '3.6'
termcolor==2.2.0 ; python_version >= '3.7'
tqdm==4.65.0 ; python_version >= '3.7'
trio==0.22.0 ; python_version >= '3.7'
trio-websocket==0.10.2 ; python_version >= '3.7'
tweepy==3.10.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
tzdata==2023.3 ; platform_system == 'Windows'
tzlocal==4.3
uiautomation==2.0.18 ; sys_platform == 'win32'
urllib3[socks]==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
validators==0.20.0 ; python_version >= '3.4'
webdriver-manager==3.8.5 ; python_version >= '3.7'
wrapt==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
xlrd==2.0.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
